backend-service-py/pyproject.toml
[project]
name = "memoria-backend"
version = "0.1.0"
description = "Memoria AI 回忆录后端服务"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web 框架
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    
    # 数据库
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    
    # 数据验证
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    
    # 认证（FastAPI 官方推荐）
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    
    # 缓存/存储
    "redis>=5.0.1",
    "minio>=7.2.3",
    
    # HTTP 客户端
    "httpx>=0.26.0",
    "websockets>=12.0",
    
    # 工具
    "python-dotenv>=1.0.0",
    "nanoid>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
